moved some components out of the page-name.js so that they could be edited and all article pages will update
