+++ src/pages/blog/template.js
@@ -3,13 +3,19 @@ import { Wrapper } from "../../components/Markdown-Wrapper"
 import "../../components/blog.css"
 import { createShortUrl, preloadUrls } from "../../utils/urlShortener"
 import { getShareCount, incrementShareCount, formatShareCount, setMockShareCount } from "../../utils/shareCounter"
+import ShareButton from "../../components/ShareButton"
+import TableOfContents from "../../components/TableOfContents"
+import ArticleHeader from "../../components/ArticleHeader"
+import ActionBar from "../../components/ActionBar"
+import RelatedArticles from "../../components/RelatedArticles"
 
 // ============================================================
 // ARTICLE METADATA - EDIT THIS SECTION FOR EACH NEW ARTICLE
 // ============================================================
 const articleMetadata = {
   title: "JavaScript Blog Template Guide - How to Create Articles",
-  description: "Complete guide for developers on how to use the JavaScript template system for creating KarmaCall blog articles. Includes metadata setup, content structure, and advanced features.",
+  description:
+    "Complete guide for developers on how to use the JavaScript template system for creating KarmaCall blog articles. Includes metadata setup, content structure, and advanced features.",
   author: "Draven Grondona",
   date: "2025-06-07", // Format: YYYY-MM-DD
   featuredImage: "../../images/blog/your-image-filename.jpg", // Place image in src/images/blog/
@@ -23,36 +29,8 @@ const articleMetadata = {
 // MAIN COMPONENT
 // ============================================================
 export default function Template() {
-  const [showShareDialog, setShowShareDialog] = React.useState(false)
-  const [linkCopied, setLinkCopied] = React.useState(false)
   const [shareCount, setShareCount] = React.useState(0)
-  const [tocCollapsed, setTocCollapsed] = React.useState(false)
   const [showScrollToTop, setShowScrollToTop] = React.useState(false)
-  const [headings, setHeadings] = React.useState([])
-
-  const handleShare = () => {
-    setShowShareDialog(!showShareDialog)
-  }
-
-  const handleCloseShare = () => {
-    setShowShareDialog(false)
-    setLinkCopied(false) // Reset copied state when dialog closes
-  }
-
-  const toggleToc = () => {
-    setTocCollapsed(!tocCollapsed)
-  }
-
-  const handleSmoothScroll = (e, targetId) => {
-    e.preventDefault()
-    const targetElement = document.getElementById(targetId)
-    if (targetElement) {
-      targetElement.scrollIntoView({
-        behavior: "smooth",
-        block: "start",
-      })
-    }
-  }
 
   const scrollToTop = () => {
     window.scrollTo({
@@ -60,56 +38,6 @@ export default function Template() {
       behavior: "smooth",
     })
   }
-
-  const handleShareAction = platform => {
-    if (typeof window !== "undefined") {
-      const currentPath = window.location.pathname
-      const newCount = incrementShareCount(currentPath)
-      setShareCount(newCount)
-      console.log(`📈 Share count incremented to ${newCount} for ${platform}`)
-    }
-  }
-
-  // Generate table of contents from headings
-  React.useEffect(() => {
-    const generateTOC = () => {
-      if (typeof window !== "undefined") {
-        // Find all H2 and H3 elements with IDs in document order
-        const allHeadings = Array.from(document.querySelectorAll("h2[id], h3[id]"))
-
-        const headingsList = []
-        let currentH2 = null
-
-        allHeadings.forEach(heading => {
-          if (heading.tagName === "H2") {
-            // Start a new H2 section
-            currentH2 = {
-              id: heading.id,
-              text: heading.textContent,
-              level: 2,
-              children: [],
-            }
-            headingsList.push(currentH2)
-          } else if (heading.tagName === "H3" && currentH2) {
-            // Add H3 as child of current H2
-            currentH2.children.push({
-              id: heading.id,
-              text: heading.textContent,
-              level: 3,
-            })
-          }
-        })
-
-        setHeadings(headingsList)
-        console.log("📋 Generated table of contents:", headingsList)
-      }
-    }
-
-    // Generate TOC after component mounts and content is rendered
-    const timer = setTimeout(generateTOC, 100)
-
-    return () => clearTimeout(timer)
-  }, [])
 
   // Preload URLs and load share count when component mounts
   React.useEffect(() => {
@@ -265,190 +193,6 @@ export default function Template() {
     }
   }, [])
 
-  const copyToClipboard = () => {
-    if (typeof window !== "undefined" && navigator.clipboard) {
-      const currentUrl = window.location.href.split("?")[0] // Remove existing params
-      const shortUrl = createShortUrl(currentUrl, "copy_link")
-      navigator.clipboard.writeText(shortUrl)
-      setLinkCopied(true)
-
-      // Increment share count
-      handleShareAction("copy_link")
-
-      // Get stored UTM data for enhanced tracking
-      const storedUTM = sessionStorage.getItem("utm_data")
-      const utmData = storedUTM ? JSON.parse(storedUTM) : {}
-
-      // Track with Google Analytics
-      if (typeof window !== "undefined" && window.gtag) {
-        window.gtag("event", "share", {
-          method: "copy_link",
-          content_type: "blog_post",
-          item_id: window.location.pathname,
-          // Include original UTM data to track full journey
-          original_utm_source: utmData.source || null,
-          original_utm_medium: utmData.medium || null,
-          original_utm_campaign: utmData.campaign || null,
-        })
-      }
-    }
-  }
-
-  const shareOptions = [
-    {
-      name: linkCopied ? "Link Copied" : "Copy Link",
-      icon: "🔗",
-      action: copyToClipboard,
-      isSuccess: linkCopied,
-    },
-    {
-      name: "Email",
-      icon: "📧",
-      action: () => {
-        if (typeof window !== "undefined") {
-          const currentUrl = window.location.href.split("?")[0]
-          const shortUrl = createShortUrl(currentUrl, "email")
-          window.open(`mailto:?subject=Check out this article&body=${shortUrl}`)
-
-          // Increment share count
-          handleShareAction("email")
-
-          // Get stored UTM data for enhanced tracking
-          const storedUTM = sessionStorage.getItem("utm_data")
-          const utmData = storedUTM ? JSON.parse(storedUTM) : {}
-
-          // Track with Google Analytics
-          if (window.gtag) {
-            window.gtag("event", "share", {
-              method: "email",
-              content_type: "blog_post",
-              item_id: window.location.pathname,
-              // Include original UTM data to track full journey
-              original_utm_source: utmData.source || null,
-              original_utm_medium: utmData.medium || null,
-              original_utm_campaign: utmData.campaign || null,
-            })
-          }
-        }
-      },
-    },
-    {
-      name: "Facebook",
-      icon: "📘",
-      action: () => {
-        if (typeof window !== "undefined") {
-          const currentUrl = window.location.href.split("?")[0]
-          const shortUrl = createShortUrl(currentUrl, "facebook")
-          window.open(`https://www.facebook.com/sharer/sharer.php?u=${encodeURIComponent(shortUrl)}`)
-
-          // Increment share count
-          handleShareAction("facebook")
-
-          // Track with Google Analytics
-          if (window.gtag) {
-            window.gtag("event", "share", {
-              method: "facebook",
-              content_type: "blog_post",
-              item_id: window.location.pathname,
-            })
-          }
-        }
-      },
-    },
-    {
-      name: "X (Twitter)",
-      icon: "🐦",
-      action: () => {
-        if (typeof window !== "undefined") {
-          const currentUrl = window.location.href.split("?")[0]
-          const shortUrl = createShortUrl(currentUrl, "twitter")
-          window.open(`https://twitter.com/intent/tweet?url=${encodeURIComponent(shortUrl)}&text=Check out this article`)
-
-          // Increment share count
-          handleShareAction("twitter")
-
-          // Track with Google Analytics
-          if (window.gtag) {
-            window.gtag("event", "share", {
-              method: "twitter",
-              content_type: "blog_post",
-              item_id: window.location.pathname,
-            })
-          }
-        }
-      },
-    },
-    {
-      name: "LinkedIn",
-      icon: "💼",
-      action: () => {
-        if (typeof window !== "undefined") {
-          const currentUrl = window.location.href.split("?")[0]
-          const shortUrl = createShortUrl(currentUrl, "linkedin")
-          window.open(`https://www.linkedin.com/sharing/share-offsite/?url=${encodeURIComponent(shortUrl)}`)
-
-          // Increment share count
-          handleShareAction("linkedin")
-
-          // Track with Google Analytics
-          if (window.gtag) {
-            window.gtag("event", "share", {
-              method: "linkedin",
-              content_type: "blog_post",
-              item_id: window.location.pathname,
-            })
-          }
-        }
-      },
-    },
-    {
-      name: "Reddit",
-      icon: "🤖",
-      action: () => {
-        if (typeof window !== "undefined") {
-          const currentUrl = window.location.href.split("?")[0]
-          const shortUrl = createShortUrl(currentUrl, "reddit")
-          window.open(`https://reddit.com/submit?url=${encodeURIComponent(shortUrl)}&title=Check out this article`)
-
-          // Increment share count
-          handleShareAction("reddit")
-
-          // Track with Google Analytics
-          if (window.gtag) {
-            window.gtag("event", "share", {
-              method: "reddit",
-              content_type: "blog_post",
-              item_id: window.location.pathname,
-            })
-          }
-        }
-      },
-    },
-    {
-      name: "Bluesky",
-      icon: "🦋",
-      action: () => {
-        if (typeof window !== "undefined") {
-          const currentUrl = window.location.href.split("?")[0]
-          const shortUrl = createShortUrl(currentUrl, "bluesky")
-          window.open(`https://bsky.app/intent/compose?text=Check out this article: ${encodeURIComponent(shortUrl)}`)
-
-          // Increment share count
-          handleShareAction("bluesky")
-
-          // Track with Google Analytics
-          if (window.gtag) {
-            window.gtag("event", "share", {
-              method: "bluesky",
-              content_type: "blog_post",
-              item_id: window.location.pathname,
-            })
-          }
-        }
-      },
-    },
-  ]
-
   // This page is completely removed in production builds via gatsby-node.js
   // This is just an extra safety check
   if (typeof window !== "undefined" && window.location.hostname !== "localhost") {
@@ -465,223 +209,15 @@ export default function Template() {
   return (
     <Wrapper seo={seo}>
       <div style={{ maxWidth: "1200px", margin: "0 auto", padding: "2rem" }}>
-        {/* Header section with same width as main content */}
-        <div style={{ display: "flex", gap: "3rem", alignItems: "flex-start", marginBottom: "1.5rem" }}>
-          {/* Header content - same width as main content column */}
-          <div style={{ flex: "1", minWidth: "0" }}>
-            {/* Title */}
-            <h1
-              style={{
-                fontSize: "2.5rem",
-                fontWeight: "700",
-                lineHeight: "1.2",
-                marginBottom: "1rem",
-                color: "var(--color-text, #333)",
-              }}
-            >
-              {articleMetadata.title}
-            </h1>
-
-            {/* Meta information */}
-            <div
-              className="blog-meta"
-              style={{
-                marginBottom: "2rem",
-                fontSize: "1rem",
-                borderBottom: "1px solid var(--border-color, #eee)",
-                paddingBottom: "1rem",
-              }}
-            >
-              <span className="blog-author">{articleMetadata.author}</span>
-              <span className="blog-date">
-                {new Date(articleMetadata.date).toLocaleDateString("en-US", {
-                  year: "numeric",
-                  month: "long",
-                  day: "numeric",
-                })}
-              </span>
-            </div>
-
-            {/* Article summary/blurb */}
-            <div
-              style={{
-                marginBottom: "0",
-              }}
-            >
-              <p
-                style={{
-                  fontSize: "1.2rem",
-                  lineHeight: "1.6",
-                  color: "var(--color-text-secondary, #666)",
-                  fontStyle: "italic",
-                  marginBottom: "0",
-                }}
-              >
-                {articleMetadata.description}
-              </p>
-            </div>
-          </div>
-
-          {/* Empty space to match sidebar width */}
-          <div
-            style={{
-              width: "280px",
-              flexShrink: 0,
-            }}
-          >
-            {/* This empty div ensures the header content aligns with main content */}
-          </div>
-        </div>
+        {/* Article Header */}
+        <ArticleHeader articleData={articleMetadata} />
 
         {/* Main content container with sidebar layout - starts at featured image level */}
         <div style={{ display: "flex", gap: "3rem", alignItems: "flex-start", marginTop: "0.5rem", position: "relative" }}>
           {/* Main article content */}
           <div style={{ flex: "1", minWidth: "0" }}>
-            {/* Action bar above featured image */}
-            <div
-              style={{
-                display: "flex",
-                justifyContent: "flex-end",
-                alignItems: "center",
-                gap: "0.1rem",
-                marginBottom: "0.75rem",
-                position: "relative",
-              }}
-            >
-              <button
-                onClick={handleShare}
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                  gap: formatShareCount(shareCount) ? "0.3rem" : "0",
-                  width: formatShareCount(shareCount) ? "auto" : "32px",
-                  height: "32px",
-                  padding: formatShareCount(shareCount) ? "0 0.6rem" : "0",
-                  backgroundColor: showShareDialog ? "lightblue" : "transparent",
-                  border: "2px solid var(--border-color, #ddd)",
-                  borderRadius: formatShareCount(shareCount) ? "16px" : "50%",
-                  color: "var(--color-text, #333)",
-                  fontSize: "0.9rem",
-                  cursor: "pointer",
-                  transition: "all 0.2s ease",
-                  whiteSpace: "nowrap",
-                }}
-              >
-                <span>📤</span>
-                {formatShareCount(shareCount) && <span style={{ fontSize: "0.8rem", fontWeight: "600" }}>{formatShareCount(shareCount)}</span>}
-              </button>
-
-              <button
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                  width: "32px",
-                  height: "32px",
-                  backgroundColor: "transparent",
-                  border: "2px solid var(--border-color, #ddd)",
-                  borderRadius: "50%",
-                  color: "var(--color-text, #333)",
-                  fontSize: "0.9rem",
-                  cursor: "pointer",
-                  transition: "all 0.2s ease",
-                }}
-              >
-                💬
-              </button>
-
-              {/* Share Dialog - positioned directly under share button with right sides aligned */}
-              {showShareDialog && (
-                <>
-                  {/* Invisible overlay to catch clicks outside the dialog */}
-                  <div
-                    style={{
-                      position: "fixed",
-                      top: 0,
-                      left: 0,
-                      right: 0,
-                      bottom: 0,
-                      zIndex: 999,
-                    }}
-                    onClick={handleCloseShare}
-                  />
-                  {/* The actual dialog */}
-                  <div
-                    style={{
-                      position: "absolute",
-                      top: "calc(100% + 0.5rem)",
-                      right: "32px",
-                      backgroundColor: "var(--color-background, white)",
-                      borderRadius: "8px",
-                      padding: "0.75rem",
-                      border: "1px solid var(--border-color, #eee)",
-                      boxShadow: "0 4px 12px rgba(0, 0, 0, 0.15)",
-                      zIndex: 1000,
-                      whiteSpace: "nowrap",
-                    }}
-                    onClick={e => e.stopPropagation()}
-                  >
-                    {/* Share options in single row */}
-                    <div
-                      style={{
-                        display: "flex",
-                        alignItems: "flex-start",
-                        gap: "0",
-                      }}
-                    >
-                      {shareOptions.map((option, index) => (
-                        <button
-                          key={index}
-                          onClick={option.action}
-                          style={{
-                            display: "flex",
-                            flexDirection: "column",
-                            alignItems: "center",
-                            gap: "0.15rem",
-                            padding: "0",
-                            margin: "0",
-                            backgroundColor: "transparent",
-                            border: "none",
-                            cursor: "pointer",
-                            transition: "all 0.2s ease",
-                            minWidth: "50px",
-                            width: "50px",
-                          }}
-                        >
-                          <div
-                            style={{
-                              width: "32px",
-                              height: "32px",
-                              borderRadius: "50%",
-                              backgroundColor: "var(--color-background-alt, #f9f9f9)",
-                              border: "1px solid var(--border-color, #eee)",
-                              display: "flex",
-                              alignItems: "center",
-                              justifyContent: "center",
-                              fontSize: "1rem",
-                              transition: "all 0.2s ease",
-                            }}
-                          >
-                            {option.icon}
-                          </div>
-                          <span
-                            style={{
-                              fontSize: "0.6rem",
-                              color: option.isSuccess ? "#28a745" : "var(--color-text, #333)",
-                              textAlign: "center",
-                              lineHeight: "1",
-                            }}
-                          >
-                            {option.name}
-                          </span>
-                        </button>
-                      ))}
-                    </div>
-                  </div>
-                </>
-              )}
-            </div>
+            {/* Action bar with share and comment buttons */}
+            <ActionBar articleData={articleMetadata} shareCount={shareCount} onShareCountUpdate={setShareCount} />
 
             {/* Featured image */}
             <div
@@ -1098,345 +634,11 @@ export default function Template() {
             </div>
 
             {/* Related Articles Section */}
-            <div
-              style={{
-                marginTop: "4rem",
-                paddingTop: "2rem",
-                borderTop: "2px solid var(--border-color, #eee)",
-              }}
-            >
-              <h2
-                style={{
-                  fontSize: "1.8rem",
-                  fontWeight: "600",
-                  marginBottom: "2rem",
-                  color: "var(--color-text, #333)",
-                  textAlign: "center",
-                }}
-              >
-                Related Articles
-              </h2>
-
-              <div
-                style={{
-                  display: "grid",
-                  gridTemplateColumns: "repeat(auto-fit, minmax(250px, 1fr))",
-                  gap: "1.5rem",
-                  marginBottom: "2rem",
-                }}
-              >
-                {/* Article 1 - Future of Spam Blocking */}
-                <a
-                  href="/blog/future-of-spam-blocking?utm_source=blog_article&utm_medium=related_articles&utm_campaign=internal_linking"
-                  style={{
-                    textDecoration: "none",
-                    color: "inherit",
-                    display: "block",
-                    border: "1px solid var(--border-color, #eee)",
-                    borderRadius: "8px",
-                    overflow: "hidden",
-                    transition: "all 0.2s ease",
-                    backgroundColor: "var(--color-background, white)",
-                  }}
-                  onMouseEnter={e => {
-                    e.target.style.transform = "translateY(-4px)"
-                    e.target.style.boxShadow = "0 8px 24px rgba(0, 0, 0, 0.12)"
-                  }}
-                  onMouseLeave={e => {
-                    e.target.style.transform = "translateY(0)"
-                    e.target.style.boxShadow = "0 2px 8px rgba(0, 0, 0, 0.08)"
-                  }}
-                >
-                  <div
-                    style={{
-                      height: "140px",
-                      backgroundColor: "var(--color-background-alt, #f9f9f9)",
-                      backgroundImage: "url('../../images/blog/interactive-rewards-blog-social-graphic.jpg')",
-                      backgroundSize: "cover",
-                      backgroundPosition: "center",
-                    }}
-                  />
-                  <div style={{ padding: "1rem" }}>
-                    <h3
-                      style={{
-                        fontSize: "1.1rem",
-                        fontWeight: "600",
-                        marginBottom: "0.5rem",
-                        color: "var(--color-text, #333)",
-                        lineHeight: "1.3",
-                      }}
-                    >
-                      Get Cash Back for Blocking Spam, with KarmaCall Version 4.0
-                    </h3>
-                    <p
-                      style={{
-                        fontSize: "0.85rem",
-                        color: "var(--color-text-secondary, #666)",
-                        lineHeight: "1.4",
-                        margin: "0 0 0.75rem 0",
-                      }}
-                    >
-                      KarmaCall 4.0 is a revolutionary new app that pays you to block spam calls. With its fresh new UI and infinitely long call blocking
-                      capability...
-                    </p>
-                    <div
-                      style={{
-                        fontSize: "0.75rem",
-                        color: "var(--color-text-secondary, #666)",
-                      }}
-                    >
-                      March 11, 2024 • KarmaCall Team
-                    </div>
-                  </div>
-                </a>
-
-                {/* Article 2 - Job Scam Texts */}
-                <a
-                  href="/blog/job-scam-texts-surge-2024?utm_source=blog_article&utm_medium=related_articles&utm_campaign=internal_linking"
-                  style={{
-                    textDecoration: "none",
-                    color: "inherit",
-                    display: "block",
-                    border: "1px solid var(--border-color, #eee)",
-                    borderRadius: "8px",
-                    overflow: "hidden",
-                    transition: "all 0.2s ease",
-                    backgroundColor: "var(--color-background, white)",
-                  }}
-                  onMouseEnter={e => {
-                    e.target.style.transform = "translateY(-4px)"
-                    e.target.style.boxShadow = "0 8px 24px rgba(0, 0, 0, 0.12)"
-                  }}
-                  onMouseLeave={e => {
-                    e.target.style.transform = "translateY(0)"
-                    e.target.style.boxShadow = "0 2px 8px rgba(0, 0, 0, 0.08)"
-                  }}
-                >
-                  <div
-                    style={{
-                      height: "140px",
-                      backgroundColor: "var(--color-background-alt, #f9f9f9)",
-                      backgroundImage: "url('../../images/illustrations/inbox-money.png')",
-                      backgroundSize: "cover",
-                      backgroundPosition: "center",
-                    }}
-                  />
-                  <div style={{ padding: "1rem" }}>
-                    <h3
-                      style={{
-                        fontSize: "1.1rem",
-                        fontWeight: "600",
-                        marginBottom: "0.5rem",
-                        color: "var(--color-text, #333)",
-                        lineHeight: "1.3",
-                      }}
-                    >
-                      Job Scam Texts Cost Americans $470M in 2024 - Here's the Economic Solution
-                    </h3>
-                    <p
-                      style={{
-                        fontSize: "0.85rem",
-                        color: "var(--color-text-secondary, #666)",
-                        lineHeight: "1.4",
-                        margin: "0 0 0.75rem 0",
-                      }}
-                    >
-                      Job scam texts were the #2 most common hoax in 2024, costing Americans nearly half a billion dollars. Discover how FynCom's refundable
-                      deposit technology...
-                    </p>
-                    <div
-                      style={{
-                        fontSize: "0.75rem",
-                        color: "var(--color-text-secondary, #666)",
-                      }}
-                    >
-                      June 7, 2024 • KarmaCall Team
-                    </div>
-                  </div>
-                </a>
-
-                {/* Article 3 - Placeholder */}
-                <div
-                  style={{
-                    border: "2px dashed var(--border-color, #eee)",
-                    borderRadius: "8px",
-                    padding: "1.5rem",
-                    textAlign: "center",
-                    backgroundColor: "var(--color-background-alt, #f9f9f9)",
-                  }}
-                >
-                  <div
-                    style={{
-                      fontSize: "1.8rem",
-                      marginBottom: "0.75rem",
-                      opacity: "0.5",
-                    }}
-                  >
-                    📝
-                  </div>
-                  <h3
-                    style={{
-                      fontSize: "1.1rem",
-                      fontWeight: "600",
-                      marginBottom: "0.5rem",
-                      color: "var(--color-text-secondary, #666)",
-                    }}
-                  >
-                    Coming Soon
-                  </h3>
-                  <p
-                    style={{
-                      fontSize: "0.85rem",
-                      color: "var(--color-text-secondary, #666)",
-                      margin: "0",
-                    }}
-                  >
-                    More insightful articles about communication technology and digital privacy coming soon.
-                  </p>
-                </div>
-              </div>
-            </div>
+            <RelatedArticles currentArticleSlug={articleMetadata.slug} />
           </div>
 
           {/* Table of Contents Sidebar */}
-          <div
-            style={{
-              width: "280px",
-              flexShrink: 0,
-              position: "sticky",
-              top: "2rem",
-              alignSelf: "flex-start",
-            }}
-          >
-            <div
-              style={{
-                backgroundColor: "var(--color-background-alt, #f9f9f9)",
-                border: "1px solid var(--border-color, #eee)",
-                borderRadius: "8px",
-                padding: "1.5rem",
-              }}
-            >
-              <div
-                onClick={toggleToc}
-                style={{
-                  display: "flex",
-                  justifyContent: "space-between",
-                  alignItems: "center",
-                  marginBottom: tocCollapsed ? "0" : "1rem",
-                  position: "relative",
-                  cursor: "pointer",
-                  marginLeft: "-1.5rem",
-                  marginRight: "-1.5rem",
-                  paddingLeft: "1.5rem",
-                  paddingRight: "1.5rem",
-                }}
-              >
-                <h3
-                  style={{
-                    fontSize: "1.1rem",
-                    fontWeight: "700",
-                    margin: "0",
-                    color: "white",
-                    pointerEvents: "none",
-                    lineHeight: "1.3",
-                  }}
-                >
-                  {articleMetadata.title}
-                </h3>
-                <div
-                  style={{
-                    fontSize: "0.6rem",
-                    color: "var(--color-text, #333)",
-                    transition: "transform 0.2s ease",
-                    transform: tocCollapsed ? "rotate(90deg)" : "rotate(0deg)",
-                    pointerEvents: "none",
-                    lineHeight: "1",
-                  }}
-                >
-                  ▼
-                </div>
-              </div>
-
-              {!tocCollapsed && (
-                <nav>
-                  <ul
-                    style={{
-                      listStyle: "none",
-                      padding: 0,
-                      margin: 0,
-                    }}
-                  >
-                    {headings.map((heading, index) => (
-                      <li key={heading.id} style={{ marginBottom: "0.5rem" }}>
-                        <a
-                          href={`#${heading.id}`}
-                          onClick={e => handleSmoothScroll(e, heading.id)}
-                          style={{
-                            color: "white",
-                            textDecoration: "none",
-                            fontSize: "0.95rem",
-                            fontWeight: "600",
-                            lineHeight: "1.4",
-                            display: "block",
-                            padding: "0.25rem 0",
-                            borderLeft: "3px solid transparent",
-                            paddingLeft: "0.75rem",
-                            transition: "all 0.2s ease",
-                          }}
-                        >
-                          {heading.text}
-                        </a>
-                        {heading.children.length > 0 && (
-                          <ul
-                            style={{
-                              listStyle: "none",
-                              padding: 0,
-                              margin: "0.5rem 0 0 1rem",
-                            }}
-                          >
-                            {heading.children.map((child, childIndex) => (
-                              <li key={child.id} style={{ marginBottom: "0.25rem" }}>
-                                <a
-                                  href={`#${child.id}`}
-                                  onClick={e => handleSmoothScroll(e, child.id)}
-                                  style={{
-                                    color: "#e0e0e0",
-                                    textDecoration: "none",
-                                    fontSize: "0.85rem",
-                                    fontWeight: "500",
-                                    lineHeight: "1.4",
-                                    display: "block",
-                                    padding: "0.25rem 0",
-                                    paddingLeft: "0.75rem",
-                                    transition: "all 0.2s ease",
-                                  }}
-                                >
-                                  {child.text}
-                                </a>
-                              </li>
-                            ))}
-                          </ul>
-                        )}
-                      </li>
-                    ))}
-                    {headings.length === 0 && (
-                      <li
-                        style={{
-                          padding: "1rem",
-                          textAlign: "center",
-                          color: "var(--color-text-secondary, #666)",
-                          fontSize: "0.9rem",
-                          fontStyle: "italic",
-                        }}
-                      >
-                        No headings found. Add H2 or H3 elements with id attributes to generate the table of contents.
-                      </li>
-                    )}
-                  </ul>
-                </nav>
-              )}
-            </div>
-          </div>
+          <TableOfContents title={articleMetadata.title} />
         </div>
 
         {/* Scroll-to-top button */}
